+++ src/Moq/Interception/InterceptionAspects.cs
@@ -155,7 +155,6 @@ namespace Moq
 		{
 			if (AmbientObserver.IsActive(out var observer))
 			{
-				observer.OnInvocation(mock, invocation);
 				return InterceptionAction.Continue;
 			}